Add missing testing gavs
