--- examples/component-service-example/src/main/java/io/ap4k/examples/component/Controller.java
@@ -1,13 +1,0 @@
-package io.ap4k.examples.component;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class Controller {
-
-    @RequestMapping("/")
-    public String hello() {
-        return "Hello world";
-    }
-}
